feat: week 자료형 Long -> int로 변경
Jira: S11P31A210-377

+++ backend/src/main/java/com/reviewping/coflo/domain/userproject/entity/UserProjectScore.java
@@ -24,7 +24,7 @@ public class UserProjectScore extends BaseTimeEntity {
     @JoinColumn(name = "code_quality_code_id")
     private CodeQualityCode codeQualityCode;
 
-    private Long week;
+    private int week;
     private Long totalScore;
 
     @Builder
